added final comments in Stegano.h

+++ Stegano.h
@@ -1,6 +1,20 @@
+/** @file Stegano.h
+ *  @brief Encodes and decodes images using other images.
+ * 
+ *  There are two options in this class. To encode an image or decode one. If the user wants to decode
+ *  an image, then the class takes as input an integer that represents the number of bits that are going
+ *  to be encoded in an image, it takes as imput a secret image that is going to be encrypted in a 
+ *  cover/mask images and cover/mask image where the secret image is going to be encrypted in. The
+ *  second option is to decrypt an encrypted image and get a secret image. In this case, the class takes
+ *  as input an integer that represents the number of bits that are going to be decoded in the cover/mask
+ *  image and the mask/cover image where the secret image is encrypted in.
+ *
+ *  @author AnnaMaria Sofroniou
+ *  @bug No known bugs.
+ */
+
 #ifndef STEGANO_h
 #define STEGANO_h
-
 
 /**
  * @brief Creates a mask based on the nbBits integer.
@@ -8,7 +22,7 @@
  * @param nbBits The number of bit needed for the mask.
  * @return byte The new mask.
  */
-byte findMaskByte(int nbBits);
+int findMaskByte(int nbBits);
 
 /**
  * @brief Encodes a mask image using a secret image. Bits from the secret
